apps/analytics/main.py
from fastapi import FastAPI, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
from typing import List, Optional
import pandas as pd
from datetime import datetime, timedelta
import os
from dotenv import load_dotenv

load_dotenv()

app = FastAPI(title="Donation Analytics Service", version="1.0.0")

# CORS middleware
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)


class DonationStats(BaseModel):
    total_donations: int
    total_amount: float
    average_donation: float
    unique_donors: int
    active_campaigns: int


class TrendData(BaseModel):
    date: str
    amount: float
    count: int


@app.get("/")
async def root():
    return {
        "service": "Analytics Service",
        "status": "healthy",
        "version": "1.0.0"
    }


@app.get("/health")
async def health_check():
    return {"status": "healthy", "timestamp": datetime.utcnow().isoformat()}


@app.get("/api/stats/overview", response_model=DonationStats)
async def get_overview_stats():
    """Get overall donation statistics"""
    # TODO: Fetch real data from database
    return DonationStats(
        total_donations=0,
        total_amount=0.0,
        average_donation=0.0,
        unique_donors=0,
        active_campaigns=0
    )


@app.get("/api/stats/trends", response_model=List[TrendData])
async def get_donation_trends(days: int = 30):
    """Get donation trends for specified number of days"""
    if days > 365:
        raise HTTPException(status_code=400, detail="Maximum 365 days allowed")

    # TODO: Fetch real data from database
    trends = []
    for i in range(days):
        date = datetime.utcnow() - timedelta(days=days - i - 1)
        trends.append(TrendData(
            date=date.strftime("%Y-%m-%d"),
            amount=0.0,
            count=0
        ))

    return trends


@app.get("/api/stats/campaigns/{campaign_id}")
async def get_campaign_stats(campaign_id: int):
    """Get statistics for specific campaign"""
    # TODO: Fetch campaign data from blockchain
    return {
        "campaign_id": campaign_id,
        "total_raised": 0.0,
        "goal_amount": 0.0,
        "percentage": 0.0,
        "donor_count": 0,
        "days_remaining": 0
    }


@app.get("/api/analytics/top-donors")
async def get_top_donors(limit: int = 10):
    """Get top donors by contribution amount"""
    # TODO: Fetch and analyze donor data
    return {"donors": []}


@app.get("/api/analytics/campaign-performance")
async def get_campaign_performance():
    """Analyze performance of all campaigns"""
    # TODO: Perform campaign analysis
    return {
        "successful_campaigns": 0,
        "ongoing_campaigns": 0,
        "failed_campaigns": 0,
        "average_success_rate": 0.0
    }


@app.get("/api/analytics/predictions")
async def get_predictions(campaign_id: Optional[int] = None):
    """Predict campaign success using ML models"""
    # TODO: Implement ML prediction model
    return {
        "predicted_success_probability": 0.0,
        "predicted_final_amount": 0.0,
        "confidence": 0.0
    }


if __name__ == "__main__":
    import uvicorn
    port = int(os.getenv("PORT", 5000))
    uvicorn.run(app, host="0.0.0.0", port=port)
