apps/analytics/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pandas==2.1.4
numpy==1.26.3
web3==6.15.0
pymongo==4.6.1
redis==5.0.1
python-dotenv==1.0.0
pydantic==2.5.3
httpx==0.26.0
scikit-learn==1.4.0
matplotlib==3.8.2
seaborn==0.13.1
